+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fix GH Actions publish entrypoints
+
 ## 3.0.0 - 2024-03-15
 ### Changed
 - Modify JavaScript client's API build URL to not encode URI component